+++ packages/twenty-server/src/engine/metadata-modules/calendar-channel/constants/calendar-channel-deleted.constant.ts
@@ -1,5 +1,1 @@
 export const CALENDAR_CHANNEL_DELETED_EVENT = 'calendarChannel_deleted';
-
-export type CalendarChannelDeletedEvent = {
-  calendarChannelId: string;
-};
