+++ packages/twenty-server/src/engine/metadata-modules/calendar-channel/types/calendar-channel-deleted.type.ts
@@ -1,0 +1,3 @@
+export type CalendarChannelDeletedEvent = {
+  calendarChannelId: string;
+};
