+++ packages/twenty-server/src/engine/metadata-modules/connected-account/constants/connected-account-deleted.constant.ts
@@ -1,6 +1,1 @@
 export const CONNECTED_ACCOUNT_DELETED_EVENT = 'connectedAccount_deleted';
-
-export type ConnectedAccountDeletedEvent = {
-  connectedAccountId: string;
-  userWorkspaceId: string;
-};
