+++ packages/twenty-server/src/engine/metadata-modules/connected-account/types/connected-account-deleted.type.ts
@@ -1,0 +1,4 @@
+export type ConnectedAccountDeletedEvent = {
+  connectedAccountId: string;
+  userWorkspaceId: string;
+};
