+++ packages/twenty-server/src/engine/metadata-modules/message-channel/constants/message-channel-deleted.constant.ts
@@ -1,5 +1,1 @@
 export const MESSAGE_CHANNEL_DELETED_EVENT = 'messageChannel_deleted';
-
-export type MessageChannelDeletedEvent = {
-  messageChannelId: string;
-};
