+++ packages/twenty-server/src/engine/metadata-modules/message-channel/types/message-channel-deleted.type.ts
@@ -1,0 +1,3 @@
+export type MessageChannelDeletedEvent = {
+  messageChannelId: string;
+};
